# Conflicts:
#	QC_DVE_CORE/Release/FixedRateCashflow.obj
#	QC_DVE_CORE/Release/IborCashflow.obj
#	QC_DVE_CORE/Release/IcpClfCashflow.obj
#	QC_DVE_CORE/Release/IcpClpCashflow.obj
#	QC_DVE_CORE/Release/InterestRateIndex.obj
#	QC_DVE_CORE/Release/Leg.obj
#	QC_DVE_CORE/Release/LegFactory.obj
#	QC_DVE_CORE/Release/QCClampedSpline.obj
#	QC_DVE_CORE/Release/QCCompoundWf.obj
#	QC_DVE_CORE/Release/QCContinousWf.obj
#	QC_DVE_CORE/Release/QCDiscountBondPayoff.obj
#	QC_DVE_CORE/Release/QCFXForward.obj
#	QC_DVE_CORE/Release/QCFactoryFunctions.obj
#	QC_DVE_CORE/Release/QCFixedRatePayoff.obj
#	QC_DVE_CORE/Release/QCFloatingRatePayoff.obj
#	QC_DVE_CORE/Release/QCHelperFunctions.obj
#	QC_DVE_CORE/Release/QCIcpClfPayoff.obj
#	QC_DVE_CORE/Release/QCIcpClpPayoff.obj
#	QC_DVE_CORE/Release/QCInterestRate.obj
#	QC_DVE_CORE/Release/QCInterestRateBasisSwap.obj
#	QC_DVE_CORE/Release/QCInterestRateLeg.obj
#	QC_DVE_CORE/Release/QCInterestRatePayoff.obj
#	QC_DVE_CORE/Release/QCInterestRatePeriodsFactory.obj
#	QC_DVE_CORE/Release/QCInterpolator.obj
#	QC_DVE_CORE/Release/QCLinearInterpolator.obj
#	QC_DVE_CORE/Release/QCLinearWf.obj
#	QC_DVE_CORE/Release/QCLogLinearInterpolator.obj
#	QC_DVE_CORE/Release/QCProjectingInterestRateCurve.obj
#	QC_DVE_CORE/Release/QCTimeDepositPayoff.obj
#	QC_DVE_CORE/Release/QCWealthFactor.obj
#	QC_DVE_CORE/Release/QCZeroCouponCurve.obj
#	QC_DVE_CORE/Release/QCZeroCouponDiscountFactorCurve.obj
#	QC_DVE_CORE/Release/QCZeroCouponInterestRateCurve.obj
#	QC_DVE_CORE/Release/QCZeroCurveBootstrappingFromRates.obj
#	QC_DVE_CORE/Release/QCZeroCurveBootstrappingFromRatesAndFixedLegs.obj
#	QC_DVE_CORE/Release/QCZeroCurveBootstrappingFromRatesAndFloatingLegsProjCurve.obj
#	QC_DVE_CORE/Release/QCZeroCurveBootstrappingFromRatesFwdsAndFixedLegs.obj
#	QC_DVE_CORE/Release/QCZeroCurveBootstrappingFromRatesFwdsAndFloatingLegs.obj
#	QC_DVE_CORE/Release/SimpleCashflow.obj
#	QC_DVE_CORE/Release/SimpleMultiCurrencyCashflow.obj
#	QC_DVE_CORE/Release/Tenor.obj
#	QC_DVE_CORE/x64/Release/QC_DVE_CORE.log
#	QC_DVE_PYBIND/Release/QC30360.obj
#	QC_DVE_PYBIND/Release/QCAct360.obj
#	QC_DVE_PYBIND/Release/QCAct365.obj
#	QC_DVE_PYBIND/Release/QCActAct.obj
#	QC_DVE_PYBIND/Release/QCCompoundWf.obj
#	QC_DVE_PYBIND/Release/QCContinousWf.obj
#	QC_DVE_PYBIND/Release/QCDate.obj
#	QC_DVE_PYBIND/Release/QCDiscountBondPayoff.obj
#	QC_DVE_PYBIND/Release/QCFXForward.obj
#	QC_DVE_PYBIND/Release/QCFactoryFunctions.obj
#	QC_DVE_PYBIND/Release/QCFixedRatePayoff.obj
#	QC_DVE_PYBIND/Release/QCFloatingRatePayoff.obj
#	QC_DVE_PYBIND/Release/QCHelperFunctions.obj
#	QC_DVE_PYBIND/Release/QCIcpClfPayoff.obj
#	QC_DVE_PYBIND/Release/QCIcpClpPayoff.obj
#	QC_DVE_PYBIND/Release/QCInterestRate.obj
#	QC_DVE_PYBIND/Release/QCInterestRateBasisSwap.obj
#	QC_DVE_PYBIND/Release/QCInterestRateLeg.obj
#	QC_DVE_PYBIND/Release/QCInterestRatePayoff.obj
#	QC_DVE_PYBIND/Release/QCInterestRatePeriodsFactory.obj
#	QC_DVE_PYBIND/Release/QCInterpolator.obj
#	QC_DVE_PYBIND/Release/QCLinearInterpolator.obj
#	QC_DVE_PYBIND/Release/QCLinearWf.obj
#	QC_DVE_PYBIND/Release/QCLogLinearInterpolator.obj
#	QC_DVE_PYBIND/Release/QCProjectingInterestRateCurve.obj
#	QC_DVE_PYBIND/Release/QCTimeDepositPayoff.obj
#	QC_DVE_PYBIND/Release/QCWealthFactor.obj
#	QC_DVE_PYBIND/Release/QCZeroCouponCurve.obj
#	QC_DVE_PYBIND/Release/QCZeroCouponDiscountFactorCurve.obj
#	QC_DVE_PYBIND/Release/QCZeroCouponInterestRateCurve.obj
#	QC_DVE_PYBIND/Release/QCZeroCurveBootstrappingFromRates.obj
#	QC_DVE_PYBIND/Release/QCZeroCurveBootstrappingFromRatesAndFixedLegs.obj
#	QC_DVE_PYBIND/Release/QCZeroCurveBootstrappingFromRatesAndFloatingLegsProjCurve.obj
#	QC_DVE_PYBIND/Release/QCZeroCurveBootstrappingFromRatesFwdsAndFixedLegs.obj
#	QC_DVE_PYBIND/Release/QCZeroCurveBootstrappingFromRatesFwdsAndFloatingLegs.obj
#	QC_DVE_PYBIND/Release/QC_DVE_PYBIND.log
#	QC_DVE_PYBIND/Release/QC_DVE_PYBIND.tlog/QC_DVE_PYBIND.write.1u.tlog
#	QC_DVE_PYBIND/Release/QC_DVE_PYBIND.tlog/cl.command.1.tlog
#	QC_DVE_PYBIND/Release/QC_DVE_PYBIND.tlog/cl.read.1.tlog
#	QC_DVE_PYBIND/Release/QC_DVE_PYBIND.tlog/cl.write.1.tlog
#	QC_DVE_PYBIND/Release/QC_DVE_PYBIND.tlog/link.read.1.tlog
#	QC_DVE_PYBIND/Release/main.obj
#	QC_DVE_PYBIND/Release/vc120.pdb
#	QC_DVE_PYBIND/x64/Release/QC_DVE_PYBIND.log

+++ QC_DVE_CORE/include/QCCurve.h
@@ -63,13 +63,13 @@ public:
 	* @param position posición del par buscado.
 	* @return valor del par buscado.
 	*/
-	pair<T, double> getValuesAt(unsigned long position);
+	pair<T, double> getValuesAt(size_t position);
 
 	/*!
 	* Método que devuelve el largo de la curva modelada.
 	@return largo de la curva.
 	*/
-	long getLength();
+	size_t getLength();
 
 	/*!
 	* Destructor
@@ -164,12 +164,12 @@ template<class T> void QCCurve<T>::setAbscissaAtWithValue(unsigned long position
 	return;
 }
 
-template<class T> pair<T, double> QCCurve<T>::getValuesAt(unsigned long position)
+template<class T> pair<T, double> QCCurve<T>::getValuesAt(size_t position)
 {
 	return _values[position];
 }
 
-template<class T> long QCCurve<T>::getLength()
+template<class T> size_t QCCurve<T>::getLength()
 {
 	return _values.size();
 }
