+++ QC_DVE_CORE/source/QCDiscountBondPayoff.cpp
@@ -22,7 +22,7 @@ QCDiscountBondPayoff::QCDiscountBondPayoff(
 
 void QCDiscountBondPayoff::_setAllRates()
 {
-	unsigned int numPeriods = _irLeg->size();
+	size_t numPeriods = _irLeg->size();
 
 	_allRates.resize(numPeriods);
 	double rate{ _rate->getValue() };
@@ -30,7 +30,7 @@ void QCDiscountBondPayoff::_setAllRates()
 	{
 		_allRates.at(i) = rate;
 	}
-	long curveLength{ _discountCurve->getLength() };
+	unsigned long long curveLength{ _discountCurve->getLength() };
 	vector<double> temp(curveLength, 0.0);
 	_allRatesDerivatives.resize(numPeriods);
 	for (unsigned int i = 0; i < numPeriods; ++i)
