+++ QC_DVE_CORE/source/QCIcpClfPayoff.cpp
@@ -24,11 +24,11 @@ double QCIcpClfPayoff::getForwardRateAt(int n)
 void QCIcpClfPayoff::_setAllRates()
 {
 	//Loopea sobre los periods de interest rate leg a partir del current period
-	int numPeriods = _irLeg->size();
+	size_t numPeriods = _irLeg->size();
 	_allRates.resize(numPeriods);
 	_forwardRates.resize(numPeriods);
 	_allRatesDerivatives.resize(numPeriods);
-	unsigned int tempCurrentPeriod;
+	size_t tempCurrentPeriod;
 	if (_currentPeriod == -1)
 	{
 		tempCurrentPeriod = 0;
@@ -37,7 +37,7 @@ void QCIcpClfPayoff::_setAllRates()
 	{
 		tempCurrentPeriod = _currentPeriod;
 	}
-	for (int i = tempCurrentPeriod; i < numPeriods; ++i)
+	for (size_t i = tempCurrentPeriod; i < numPeriods; ++i)
 	{
 		auto per = _irLeg->getPeriodAt(i);
 		//Si el fixing date es menor que _valueDate lo va a calcular con fixingData
