+++ QC_DVE_CORE/source/QCZeroCurveBootstrappingFromRates.cpp
@@ -83,7 +83,7 @@ void QCZeroCurveBootstrappingFromRates::generateCurveAndDerivatives()
 		//Ejecutar generateCurve() y guardar los valores
 		for (size_t k = 0; k < numRates; ++k)
 		{
-			_curve->setOrdinateAtWithValue(k, 0.0);
+			_curve->setOrdinateAtWithValue(static_cast<unsigned long>(k), 0.0);
 		}
 
 		generateCurve();
@@ -101,7 +101,7 @@ void QCZeroCurveBootstrappingFromRates::generateCurveAndDerivatives()
 
 		for (size_t k = 0; k < numRates; ++k)
 		{
-			_curve->setOrdinateAtWithValue(k, 0.0);
+			_curve->setOrdinateAtWithValue(static_cast<unsigned long>(k), 0.0);
 		}
 
 		generateCurve();
@@ -136,14 +136,14 @@ void QCZeroCurveBootstrappingFromRates::generateCurveAndDerivatives()
 	}
 }
 
-unsigned int QCZeroCurveBootstrappingFromRates::getCurveLength()
+size_t QCZeroCurveBootstrappingFromRates::getCurveLength()
 {
 	return _curve->getLength();
 }
 
 double QCZeroCurveBootstrappingFromRates::getRateAt(size_t index)
 {
-	return _curve->getRateAt(index);
+	return _curve->getRateAt(static_cast<long>(index));
 }
 
 void QCZeroCurveBootstrappingFromRates::calculateDerivativesAt(size_t derivativeIndex)
